Implementando NFe

+++ NSSuiteCoreCSharp.Library/src/Commons/Util.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.IO;
+using System.Threading.Tasks;
 using System.Xml;
 using System.Xml.Serialization;
 
@@ -7,31 +8,31 @@ namespace NSSuiteCoreCSharp.Library.src.Commons
 {
     public static class Util
     {
-        private static int GerarDigitoVerificador(string chave)
-        {
-            int soma = 0;
-            int restoDivisao = -1;
-            int digitoVerificador = -1;
-            int pesoMultiplicacao = 2;
+        //private static int GerarDigitoVerificador(string chave)
+        //{
+        //    int soma = 0;
+        //    int restoDivisao = -1;
+        //    int digitoVerificador = -1;
+        //    int pesoMultiplicacao = 2;
 
-            for (int i = chave.Length - 1; i != -1; i--)
-            {
-                int ch = Convert.ToInt32(chave[i].ToString());
-                soma += ch * pesoMultiplicacao;
+        //    for (int i = chave.Length - 1; i != -1; i--)
+        //    {
+        //        int ch = Convert.ToInt32(chave[i].ToString());
+        //        soma += ch * pesoMultiplicacao;
 
-                if (pesoMultiplicacao < 9)
-                    pesoMultiplicacao += 1;
-                else
-                    pesoMultiplicacao = 2;
-            }
-            restoDivisao = soma % 11;
-            if (restoDivisao == 0 || restoDivisao == 1)
-                digitoVerificador = 0;
-            else
-                digitoVerificador = 11 - restoDivisao;
+        //        if (pesoMultiplicacao < 9)
+        //            pesoMultiplicacao += 1;
+        //        else
+        //            pesoMultiplicacao = 2;
+        //    }
+        //    restoDivisao = soma % 11;
+        //    if (restoDivisao == 0 || restoDivisao == 1)
+        //        digitoVerificador = 0;
+        //    else
+        //        digitoVerificador = 11 - restoDivisao;
 
-            return digitoVerificador;
-        }
+        //    return digitoVerificador;
+        //}
         //public static string GerarChaveDocumento(Projetos projeto, string cUF, string dhEmi,
         //    string serie, string nDF, string tpEmis, string cDF, string cnpjEmitente,
         //    string tpEvento = "", string nSeqEvento = "")
@@ -47,7 +48,30 @@ namespace NSSuiteCoreCSharp.Library.src.Commons
         //    string chave = tpEvento + chave43 + GerarDigitoVerificador(chave43) + nSeqEvento;
         //    return chave;
         //}
-        public static string GerarXMLDocumento<T>(T ObjectXML)
+       
+        public static string GerarPDFTemporario(string pdf)
+        {
+            byte[] bytes = Convert.FromBase64String(pdf);
+            string file = Path.GetTempFileName() + ".pdf";
+            File.WriteAllBytes(file, bytes);
+            System.Diagnostics.Process.Start(@"cmd.exe ", @"/c "+ file);
+            return file;
+        }
+
+
+        public static string LerElementoDeXMLString(string xml, string elemento)
+        {
+            XmlDocument doc = new XmlDocument();
+            doc.LoadXml(xml);
+
+            XmlElement root = doc.DocumentElement;
+            XmlNodeList xmlNodeList = root.GetElementsByTagName(elemento);
+            if (xmlNodeList.Count > 1)
+                throw new ArgumentOutOfRangeException();
+
+            return root.GetElementsByTagName(elemento).Item(0).InnerText;
+        }        
+        public static string SerializeXMLObject<T>(T ObjectXML)
         {
             XmlDocument xmlDoc = new XmlDocument();
             XmlSerializer xmlSerializer = new XmlSerializer(ObjectXML.GetType());
@@ -62,6 +86,15 @@ namespace NSSuiteCoreCSharp.Library.src.Commons
                 return sw.ToString();
             }
         }
+
+        public static object DeserializeXMLObject<T>(string strXML) where T : class
+        {          
+            XmlSerializer ser = new XmlSerializer(typeof(T));
+            using (StringReader sr = new StringReader(strXML))
+                return (T)ser.Deserialize(sr);
+            
+        }
+
         public static string CriarDiretorio(string caminho)
         {
             try
@@ -77,24 +110,43 @@ namespace NSSuiteCoreCSharp.Library.src.Commons
                 throw new Exception("Erro: " + ex.Message);
             }
         }
-        public static void SalvarXML(string xml, string caminho, string nome)
+        public static async Task SalvarXML(string xml, string caminho, string nome)
         {
-            string localParaSalvar = $"{caminho + nome}.xml";
-            string ConteudoSalvar = xml.Replace(@"\""", "");
-            File.WriteAllText(localParaSalvar, ConteudoSalvar);
+            await Task.Run(() =>
+            {
+                if (!string.IsNullOrEmpty(xml))
+                {
+                    string localParaSalvar = $"{caminho + nome}.xml";
+                    string ConteudoSalvar = xml.Replace(@"\""", "");
+                    File.WriteAllText(localParaSalvar, ConteudoSalvar);
+                }
+                else
+                {
+                    Util.GravarLinhaLog("[ERRO AO SALVAR XML] : XML sem conteudo");
+                }
+            });
         }
-        public static void ExibePDFSemSalvar()
-        {
 
-        }
-        public static void SalvarPDF(string pdf, string caminho, string nome)
+        public static async Task SalvarPDF(string pdf, string caminho, string nome, bool exibirPDFNaTela)
         {
-            string localParaSalvar = $"{caminho + nome}.pdf";
-            byte[] bytes = Convert.FromBase64String(pdf);
-            if (File.Exists(localParaSalvar))
-                File.Delete(localParaSalvar);
-            using (BinaryWriter writer = new BinaryWriter(new FileStream(localParaSalvar, FileMode.CreateNew)))
-                writer.Write(bytes, 0, bytes.Length);
+            await Task.Run(() => {
+                if (!string.IsNullOrEmpty(pdf))
+                {                    
+                    string localParaSalvar = $"{caminho + nome}.pdf";
+                    byte[] bytes = Convert.FromBase64String(pdf);
+                    if (File.Exists(localParaSalvar))
+                        File.Delete(localParaSalvar);
+                    using (BinaryWriter writer = new BinaryWriter(new FileStream(localParaSalvar, FileMode.CreateNew)))
+                        writer.Write(bytes, 0, bytes.Length);
+
+                    if (exibirPDFNaTela)                   
+                        System.Diagnostics.Process.Start(@"cmd.exe ", @"/c " + localParaSalvar);               
+                }
+                else
+                {
+                    Util.GravarLinhaLog("[ERRO AO SALVAR PDF] : PDF sem conteudo");
+                }
+            }); 
         }
         public static void GravarLinhaLog(string conteudo)
         {
@@ -103,11 +155,13 @@ namespace NSSuiteCoreCSharp.Library.src.Commons
             if (!Directory.Exists(caminho))
                 Directory.CreateDirectory(caminho);
             string data = DateTime.Now.ToString("MM/dd/yyyy hh:mm:ss.fff");
-            string nomeArq = DateTime.Now.ToShortDateString().Replace("/", "");
+            string nomeArq = DateTime.Now.ToString("MM-dd-yyyy");
 
             using (StreamWriter outputFile = new StreamWriter($"{caminho + nomeArq}.log", true))
+            {
                 outputFile.WriteLine(data + " - " + conteudo);
-        }
+            }
 
+        }
     }
 }
